Rename packages to gatekeepersdk to be consistent with project name
