+++ app/src/main/java/co/blustor/gatekeepersdk/utils/GKFileUtils.java
@@ -1,4 +1,4 @@
-package co.blustor.gatekeeper.utils;
+package co.blustor.gatekeepersdk.utils;
 
 import android.support.annotation.NonNull;
